# packages/schemas/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateRunRequest": {
        "properties": {
          "excluded_citations": {
            "items": {
              "type": "string"
            },
            "maxItems": 20,
            "title": "Excluded Citations",
            "type": "array"
          },
          "goal": {
            "maxLength": 10000,
            "minLength": 1,
            "title": "Goal",
            "type": "string"
          },
          "origin": {
            "default": "ui",
            "enum": [
              "ui",
              "discord",
              "schedule"
            ],
            "title": "Origin",
            "type": "string"
          },
          "origin_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Origin Ref"
          },
          "space_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Space Id"
          }
        },
        "required": [
          "goal"
        ],
        "title": "CreateRunRequest",
        "type": "object"
      },
...
      "Event": {
        "description": "One append-only row of the log.\n\n``seq`` is per-run and 1-based: the SSE id and the ``Last-Event-ID``\ncursor. ``id`` is the global rowid and is never a resume cursor.",
        "properties": {
          "agent_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Id"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "seq": {
            "title": "Seq",
            "type": "integer"
          },
          "ts": {
            "format": "date-time",
            "title": "Ts",
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/EventType"
          }
        },
        "required": [
          "id",
          "run_id",
          "seq",
          "type",
          "ts"
        ],
        "title": "Event",
        "type": "object"
      },
      "EventType": {
        "description": "Every event this application can emit. The list is from \u00a74, verbatim.",
        "enum": [
          "run.started",
          "run.completed",
          "run.failed",
          "run.paused",
          "run.cancelled",
          "agent.spawned",
          "agent.thinking",
          "agent.message",
          "agent.handoff",
          "agent.completed",
          "llm.request",
          "llm.token",
          "llm.response",
          "llm.error",
          "tool.requested",
          "tool.approved",
          "tool.denied",
          "tool.called",
          "tool.result",
          "tool.error",
          "approval.requested",
          "approval.resolved",
          "budget.warning",
          "budget.exceeded",
          "channel.inbound",
          "channel.outbound"
        ],
        "title": "EventType",
        "type": "string"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Run": {
        "description": "A row of the `runs` table (\u00a74).",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "goal": {
            "title": "Goal",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "origin": {
            "enum": [
              "ui",
              "discord",
              "schedule"
            ],
            "title": "Origin",
            "type": "string"
          },
          "origin_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Origin Ref"
          },
          "space_id": {
            "title": "Space Id",
            "type": "string"
          },
          "status": {
            "enum": [
              "pending",
              "running",
              "paused",
              "completed",
              "failed",
              "cancelled"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "space_id",
          "goal",
          "status",
          "origin",
          "created_at"
        ],
        "title": "Run",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "AgentBase sidecar",
    "version": "0.4.5"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/debug/fake_run": {
      "post": {
        "description": "Start a scripted run so the event spine can be exercised without a model.",
        "operationId": "fake_run_debug_fake_run_post",
        "parameters": [
          {
            "in": "query",
            "name": "step_ms",
            "required": false,
            "schema": {
              "default": 500,
              "maximum": 5000,
              "minimum": 0,
              "title": "Step Ms",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Run"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Fake Run"
      }
    },
...
    "/runs": {
      "get": {
        "description": "Recent runs, newest first: what the run picker reads. ``origin`` keeps\nonly the runs started from the window, by a schedule, or from Discord.",
        "operationId": "list_runs_runs_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "space_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Space Id"
            }
          },
          {
            "in": "query",
            "name": "origin",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "ui",
                    "discord",
                    "schedule"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Origin"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Run"
                  },
                  "title": "Response List Runs Runs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Runs"
      },
      "post": {
        "description": "Create a run and start the orchestrator on it.\n\nReturns as soon as the row exists; the client watches the run over SSE.\nStarting it is :class:`~agentbase.orchestrator.launcher.RunLauncher`'s\njob, shared with the chat channels.",
        "operationId": "create_run_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Run"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Run"
      }
    },
    "/runs/{run_id}": {
      "delete": {
        "description": "Remove a finished run, its events and its approvals. Its spend stays.\n\nA run that has not ended is a 409: the log must never be deleted from\nunder an orchestrator still appending to it. A run this process is still\ndriving is refused even if its row just turned terminal, since the\norchestrator lets go a moment after writing the status. See\n:meth:`~agentbase.events.store.EventStore.delete_run` for what is kept.",
        "operationId": "delete_run_runs__run_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Run"
      },
      "get": {
        "operationId": "get_run_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Run"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Run"
      }
    },
    "/runs/{run_id}/cancel": {
      "post": {
        "description": "Ask a run to stop.\n\n202: the run stops at its next deadline check and writes `run.cancelled`\nitself, so the row returned here may still say `running`. A finished run,\nor one this process is not driving, is a 409.",
        "operationId": "cancel_run_runs__run_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Run"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Run"
      }
    },
    "/runs/{run_id}/events": {
      "get": {
        "description": "Server-sent events for one run, resumable via `Last-Event-ID`.\n\n``after_seq`` is the same cursor for a first connection, which cannot\ncarry the header. On a reconnect the header wins if it is further along.",
        "operationId": "stream_run_events_runs__run_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "after_seq",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "After Seq",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stream Run Events"
      }
    },
    "/runs/{run_id}/events/history": {
      "get": {
        "description": "The event log as a plain array, for the dashboard's history load, tests and `curl`.",
        "operationId": "get_run_events_runs__run_id__events_history_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "after_seq",
            "required": false,
            "schema": {
              "default": 0,
              "title": "After Seq",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Event"
                  },
                  "title": "Response Get Run Events Runs  Run Id  Events History Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Run Events"
      }
    },
...
  }
}